# backend/api.yaml
openapi: 3.0.0
paths:
...
  /api/v1/managers:
    post:
      operationId: ManagerController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateManagerDto'
      responses:
        '201':
          description: Manager created successfully
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Insufficient permissions
        '409':
          description: Conflict (email already exists)
      security:
        - JWT-auth: []
      summary: Create a new manager - Chỉ MANAGER
      tags:
        - managers
    get:
      operationId: ManagerController_findAll_v1
      parameters:
        - name: page
          required: false
          in: query
          description: Số trang (bắt đầu từ 1)
          schema:
            minimum: 1
            default: 1
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Số lượng bản ghi trên mỗi trang
          schema:
            minimum: 1
            maximum: 100
            default: 10
            example: 10
            type: number
      responses:
        '200':
          description: Paginated list of managers
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: '#/components/schemas/PaginationMetadata'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Insufficient permissions
      security:
        - JWT-auth: []
      summary: Get all managers with pagination - Chỉ MANAGER
      tags:
        - managers
  /api/v1/managers/{id}:
    get:
      operationId: ManagerController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          description: Manager ID
          schema:
            type: number
      responses:
        '200':
          description: Manager details
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Insufficient permissions
        '404':
          description: Manager not found
      security:
        - JWT-auth: []
      summary: Get manager by ID - Chỉ MANAGER
      tags:
        - managers
    patch:
      operationId: ManagerController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          description: Manager ID
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateManagerDto'
      responses:
        '200':
          description: Manager updated successfully
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Insufficient permissions
        '404':
          description: Manager not found
        '409':
          description: Conflict (email already exists)
      security:
        - JWT-auth: []
      summary: Update manager - Chỉ MANAGER
      tags:
        - managers
    delete:
      operationId: ManagerController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          description: Manager ID
          schema:
            type: number
      responses:
        '204':
          description: Manager deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Insufficient permissions
        '404':
          description: Manager not found
      security:
        - JWT-auth: []
      summary: Delete manager - Chỉ MANAGER
      tags:
        - managers
  /api/v1/managers/email/{email}:
    get:
      operationId: ManagerController_findByEmail_v1
      parameters:
        - name: email
          required: true
          in: path
          description: Manager email
          schema:
            type: string
      responses:
        '200':
          description: Manager details
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Insufficient permissions
        '404':
          description: Manager not found
      security:
        - JWT-auth: []
      summary: Find manager by email - Chỉ MANAGER
      tags:
        - managers
  /api/v1/managers/{id}/account/lock:
    patch:
      description: Thay đổi trạng thái khóa của tài khoản manager
      operationId: ManagerController_lockManagerAccount_v1
      parameters:
        - name: id
          required: true
          in: path
          description: Manager ID
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockAccountDto'
      responses:
        '200':
          description: Thay đổi trạng thái khóa tài khoản thành công
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Insufficient permissions
        '404':
          description: Manager not found
      security:
        - JWT-auth: []
      summary: Khóa/mở khóa tài khoản manager - Chỉ MANAGER
      tags:
        - managers
  /api/v1/managers/{id}/account/{accountId}:
    patch:
      description: Cập nhật username, password, và trạng thái tài khoản của manager
      operationId: ManagerController_updateManagerAccount_v1
      parameters:
        - name: id
          required: true
          in: path
          description: Manager ID
          schema:
            type: number
        - name: accountId
          required: true
          in: path
          description: Account ID
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountDto'
      responses:
        '200':
          description: Cập nhật thông tin tài khoản thành công
        '400':
          description: Bad Request - Account không thuộc về Manager này
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Insufficient permissions
        '404':
          description: Manager hoặc Account không tồn tại
        '409':
          description: Conflict - Username đã tồn tại
      security:
        - JWT-auth: []
      summary: Cập nhật thông tin tài khoản của manager - Chỉ MANAGER
      tags:
        - managers
  /api/v1/managers/bulk:
    delete:
      operationId: ManagerController_bulkDelete_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteManagerDto'
      responses:
        '200':
          description: Bulk delete completed with results
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Insufficient permissions
      security:
        - JWT-auth: []
      summary: Bulk delete managers - Chỉ MANAGER
      tags:
        - managers
...
info:
  title: Cake POS API
  description: API documentation for Cake POS System - Hệ thống quản lý cửa hàng bánh ngọt
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
...
    PaginationMetadata:
      type: object
      properties:
        page:
          type: number
          description: Số trang hiện tại
          example: 1
          minimum: 1
        limit:
          type: number
          description: Số lượng bản ghi trên mỗi trang
          example: 10
          minimum: 1
          maximum: 100
        total:
          type: number
          description: Tổng số bản ghi
          example: 50
          minimum: 0
        totalPages:
          type: number
          description: Tổng số trang
          example: 5
          minimum: 0
        hasNext:
          type: boolean
          description: Có trang tiếp theo hay không
          example: true
        hasPrev:
          type: boolean
          description: Có trang trước hay không
          example: false
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
...
    UpdateAccountDto:
      type: object
      properties:
        username:
          type: string
          description: Tên đăng nhập mới (tối thiểu 3 ký tự)
          example: newuser123
          minLength: 3
        password:
          type: string
          description: Mật khẩu mới (tối thiểu 8 ký tự)
          example: newpassword123
          minLength: 8
        is_active:
          type: boolean
          description: Trạng thái hoạt động của tài khoản
          example: false
        is_locked:
          type: boolean
          description: Trạng thái khóa tài khoản
          example: true
...
    LockAccountDto:
      type: object
      properties:
        is_locked:
          type: boolean
          description: Trạng thái khóa tài khoản (true = khóa, false = mở khóa)
          example: true
      required:
        - is_locked
...
    CreateManagerDto:
      type: object
      properties:
        first_name:
          type: string
          description: Tên của quản lý
          example: Nguyễn Văn
          maxLength: 70
        last_name:
          type: string
          description: Họ của quản lý
          example: A
          maxLength: 70
        email:
          type: string
          description: Email của quản lý
          example: manager@banhngotnhalam.com
          maxLength: 100
        phone:
          type: string
          description: Số điện thoại của quản lý
          example: '0901234567'
          maxLength: 15
        gender:
          type: string
          description: Giới tính của quản lý
          example: MALE
          enum:
            - MALE
            - FEMALE
            - OTHER
        username:
          type: string
          description: Tên đăng nhập cho tài khoản quản lý
          example: manager01
          minLength: 3
          maxLength: 50
      required:
        - first_name
        - last_name
        - email
        - phone
        - username
    UpdateManagerDto:
      type: object
      properties:
        first_name:
          type: string
          description: Tên của quản lý
          example: Nguyễn Văn
          maxLength: 70
        last_name:
          type: string
          description: Họ của quản lý
          example: A
          maxLength: 70
        email:
          type: string
          description: Email của quản lý
          example: manager@banhngotnhalam.com
          maxLength: 100
        phone:
          type: string
          description: Số điện thoại của quản lý
          example: '0901234567'
          maxLength: 15
        gender:
          type: string
          description: Giới tính của quản lý
          example: MALE
          enum:
            - MALE
            - FEMALE
            - OTHER
        username:
          type: string
          description: Tên đăng nhập cho tài khoản quản lý
          example: manager01
          minLength: 3
          maxLength: 50
    BulkDeleteManagerDto:
      type: object
      properties:
        ids:
          description: Danh sách ID của các manager cần xóa
          example:
            - 1
            - 2
            - 3
          type: array
          items:
            type: number
      required:
        - ids
...